Add LocationFeeFactory
